requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.24.0
datasets>=2.14.0
sentencepiece>=0.1.99
protobuf>=3.20.0
vllm>=0.6.0
mistral_common>=1.8.8

# DeepSeek and model serving
# Note: deepseek-ai models are loaded via transformers from HuggingFace
# No separate package needed - models are at: deepseek-ai/deepseek-v3

# Retrieval and embeddings
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.19.0
tqdm>=4.65.0

# Evaluation and monitoring
scikit-learn>=1.3.0
scipy>=1.11.0
wandb>=0.15.0  # Optional for experiment tracking

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Symbolic verification
z3-solver>=4.12.0  # Z3 SMT solver for formal verification

# API clients
openai>=2.0.0  # OpenAI API for atomic proposition generation

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

